requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
httpx
passlib[bcrypt]